configure swagger to display groups, update the add-field controller to use a PUT request
